Read a column chunk holding more than 2 GiB of BYTE_ARRAY data
`large_string_map.brotli.parquet` in apache/parquet-testing is two rows
whose map keys are a gibibyte each. Its Brotli pages always decoded fine;
the chunk's 2 GiB of value bytes did not fit a 32-bit offset, and four
places in `parquet.encoding` raised rather than let one wrap negative.

Two changes, and neither is Arrow's `large_binary`.

**The physical buffer is 64-bit.** `PhysBuffer.offsets` is `List[Int]`,
not `List[Int32]`. It is not an Arrow buffer — it is where a decoded
column chunk's values live before assembly — and keeping it narrow was
what made the overflow happen during *decode*, where nothing downstream
can split it away. Widening it removes the four raises with it. On this
file it is the difference between a chunk that cannot be decoded and one
that can.

**The batch is cut instead.** Arrow's `binary`/`string` layout is still
32-bit, and a batch is Arrow arrays, so `read_batch` gets a third bound
next to `batch_size` and the row group: a batch's value bytes may not
outgrow an Arrow offset. `_arrow_batch_end` finds the last row that fits
by binary search over the chunk's own offsets — `read_table`'s planner
applies the same bound to a prefetched window — and takes the smallest
row *any* column can reach, because the columns of a batch are aligned
and a cut that suited only the column that overflowed would tear the
others apart. A chunk whose whole value buffer fits is dismissed on one
comparison, so nothing but an oversized chunk pays for this.

`max_batch_value_bytes` is that bound, public and defaulting to the 2 GiB
ceiling. It exists as a knob because the alternative was a two-gibibyte
fixture in the test suite.

A single row larger than the bound has no split point that helps, and
raises with the column and the byte count named. That is the gap that is
left, and the README now says so instead of the chunk-sized one.

**Evidence.** Conformance goes from

    data/     : 67 read, 3 known-unreadable, 0 failed

to

    data/     : 68 read, 2 known-unreadable, 0 failed

— the two that remain are deliberately corrupt and rejecting them is the
right answer. Reading the file is not the whole claim, so the values were
checked against an independent oracle: both rows come back as batches of
one, each key 1073741824 bytes with CRC32 261666223, and decompressing
the dictionary page and the second data page with Python's `brotli` and
hashing them with `zlib.crc32` gives 1073741824 and 261666223 for both.

Three negative controls, each rebuilt and run:

* never cut a batch by bytes — the 40-row `list<string>` fixture comes
  back as one batch instead of ten, and the unsplittable row does not
  raise;
* use the flat row-to-value mapping for a repeated leaf — the cut lands
  on the wrong rows and the fixture comes back as five batches, not ten;
* the corpus file itself, before the split existed: listed as
  known-unreadable, it now reports `UNEXPECTED PASS`, which fails the
  conformance run until the entry is removed.

+++ src/parquet/writer.mojo
@@ -737,7 +737,7 @@ def shred_flat(
             cols[leaf_i].values.bytes.extend(Span(a.values)[lo:hi])
             for k in range(n):
                 cols[leaf_i].values.offsets.append(
-                    Int32(vbase + Int(off.unsafe_load(r0 + k + 1)) - lo)
+                    vbase + Int(off.unsafe_load(r0 + k + 1)) - lo
                 )
         else:
             for k in range(n):
@@ -747,7 +747,7 @@ def shred_flat(
                 var hi = Int(off.unsafe_load(r0 + k + 1))
                 cols[leaf_i].values.bytes.extend(Span(a.values)[lo:hi])
                 cols[leaf_i].values.offsets.append(
-                    Int32(len(cols[leaf_i].values.bytes))
+                    len(cols[leaf_i].values.bytes)
                 )
         cols[leaf_i].values.count += present
         return True
